% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClussCluster.R
\name{ClussCluster}
\alias{ClussCluster}
\title{Function to perform simultaneous detection of cell types and cell-type-specific signature genes}
\usage{
ClussCluster(x, nclust = NULL, centers = NULL, ws = NULL,
  nepoch.max = 10, theta = NULL, seed = 1, nstart = 20, iter.max = 50,
  verbose = FALSE, progress = TRUE)
}
\arguments{
\item{x}{An nxp data matrix. There are n cells and p genes.}

\item{nclust}{Number of clusters desired if the cluster centers are not provided. If both are provided, nclust must equal the number of cluster \code{centers}.}

\item{centers}{A set of initial (distinct) cluster centres if the number of clusters (\code{nclust}) is null. If both are provided, the number of cluster centres must equal \code{nclust}.}

\item{ws}{One or multiple candidate tuning parameters to be evaluated and compared. Determines the sparsity of the selected genes. Should be greater than 1.}

\item{nepoch.max}{The maximum number of epochs. In one epoch, each cell will be evaluated to determine if its label needs to be updated.}

\item{theta}{Optional argument. If provided, \code{theta} are used as the initial cluster labels of the ClussCluster algorithm; if not, K-means is performed to produce starting cluster labels.}

\item{seed}{This seed is used wherever K-means is used.}

\item{nstart}{Argument passed to \code{kmeans}. It is the number of random sets used in \code{kmeans}.}

\item{iter.max}{Argument passed to \code{kmeans}. The maximum number of iterations allowed.}

\item{verbose}{Print the updates inside every epoch? If TRUE, the updates of cluster label and the value of objective function will be printed out.}

\item{progress}{Print the progress? If multiple tuning parameters are evaluated, then each tuning parameter will be printed when \code{progress=TRUE}.}
}
\description{
\code{ClussCluster} takes the single-cell transcriptome data and returns an object containing cell types and type-specific signature gene sets
}
\details{
Takes the normalized and log transformed number of reads mapped to genes (e.g., log(RPKM+1) or log(TPM+1) where RPKM stands for Reads Per Kilobase of transcript per Million mapped reads and TPM stands for transcripts per million) but NOT centered.
}
\examples{
\dontrun{
data(Hou)
hou.dat <-Hou$x
run.ft <- filter_gene(hou.dat)
hou.test <- ClussCluster(run.ft$dat.ft, nclust=3, ws=4, verbose = F)
}
}
