% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClussCluster.R
\name{print_ClussCluster}
\alias{print_ClussCluster}
\title{Prints out the results of \code{ClussCluster}}
\usage{
print_ClussCluster(object)
}
\arguments{
\item{object}{An object that is obtained by applying the ClussCluster function to the data set.}
}
\description{
Prints out the results of \code{ClussCluster}
}
