% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClussCluster.R
\name{print_ClussCluster_Gap}
\alias{print_ClussCluster_Gap}
\title{Prints out the results of \code{ClussCluster_Gap}
Prints the gap statistics and number of genes selected for each candidate tuning parameter.}
\usage{
print_ClussCluster_Gap(object)
}
\arguments{
\item{object}{An object that is obtained by applying the ClussCluster_Gap function to the data set.}
}
\description{
Prints out the results of \code{ClussCluster_Gap}
Prints the gap statistics and number of genes selected for each candidate tuning parameter.
}
