% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\docType{data}
\name{sim_dat}
\alias{sim_dat}
\title{A simulated expression data set.}
\format{A data frame with 60 cells on 200 genes.}
\usage{
data(sim_dat)
}
\value{
A simulated dataset used to demonstrate the application of \code{ClussCluster}.
}
\description{
This example data set containing expressing levels for 60 cells and 200 genes. The 60 cells belong to 4 cell types with 15 cells each. Each cell type is uniquely associated with 30 signature genes, i.e., the first cell type is associated with the first 30 genes, the second cell type is associated with the next 30 genes, so on and so forth. The remaining 80 genes show indistinct expression patterns among the four cell types and are considered as noise genes.
}
\examples{
data(sim_dat)
head(sim_dat)
}
\keyword{datasets}
