% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClussCluster.R
\name{plot_ClussCluster_Gap}
\alias{plot_ClussCluster_Gap}
\title{Plots the results of \code{ClussCluster_Gap}}
\usage{
plot_ClussCluster_Gap(object)
}
\arguments{
\item{object}{object obtained from \code{ClussCluster_Gap}()}
}
\description{
Plots the gap statistics and number of genes selected as the tuning parameter varies.
}
